fix warnings again

+++ src/LowestCommonAncestor.cpp
@@ -44,8 +44,8 @@ private:
             unsigned int as_uint32;
             float as_float;
         } data{};
-        data.as_float = (float)x + 0.5;
-        int n = 158 - (data.as_uint32 >> 23);
+        data.as_float = (float)x + 0.5f;
+        int n = 158 - (int)(data.as_uint32 >> 23);
         return n;
     }
